+++ Sources/KSCrash/llvm/ADT/Raygun_Optional.h
@@ -18,7 +18,7 @@
 
 #include "Raygun_None.h"
 #include "Raygun_AlignOf.h"
-#include "Compiler.h"
+#include "Raygun_Compiler.h"
 #include <cassert>
 #include <new>
 #include <utility>
